refactor(cli): voice CLI uses canonical load/save helpers + as_str display (D1)
- agent_voice.rs: drop local profile_path/load_profile/save_profile helpers;
  use crate::cmd::agent::{load_profile_for_edit, save_profile} so updated_at
  is stamped on every write
- VoiceId: add as_str() -> &'static str for direct display (replaces fragile
  serde_yaml_ng::to_string roundtrip)
- Fix hint string: mur agent voice <name> download (was missing <name>)
- VoiceAction::Download: bail! with unimplemented message instead of silent
  println + exit-0
- Move voice_id_from_str_roundtrips / voice_id_from_str_rejects_unknown tests
  from agent_voice.rs to mur-common::agent::voice_tests (roundtrip test also
  verifies as_str matches FromStr input)
