+++ mur-core/src/main.rs
@@ -1915,8 +1915,8 @@ async fn async_main() -> Result<()> {
                 }
                 VoiceAction::Disable => cmd::agent_voice::cmd_voice_disable(&name)?,
                 VoiceAction::Download => {
-                    println!(
-                        "voice download: model download not yet implemented (ships in D1 Task 3)"
+                    anyhow::bail!(
+                        "voice download not yet implemented; will ship in D1 Task 3"
                     );
                 }
             },
